Modified cmake installation
Now when cmake is used to install the package,
the directory will not differ if the installation
is in debug or release mode, so the installation
directory will be always ${CMAKE_LIB_DIR}, this
can be used because in non Windows system the
only avaliable method is release and in Windows
the installation performs in a good way not
differing Debug from Release like a lot of libraries.

Also, the DLL destination was moved to
${CMAKE_BIN_DIR}, that is the most approppriated
path to DLLs. Also added a max suffix to suppress
cmake version errors.